requirements.txt
alembic==0.9.2
astroid==2.0.2
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
demjson==2.2.4
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
idna==2.5
isort==4.3.4
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lxml==3.8.0
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.15.0
psycopg2==2.7.5
pylint==2.1.1
python-dateutil==2.6.0
python-dotenv==0.9.1
python-editor==1.0.3
requests==2.18.2
six==1.11.0
SQLAlchemy==1.1.10
typed-ast==1.1.0
typing==3.6.4
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
